Move websocket_log_session to a separate source file
